Move decorators to their own submodule
